hook login for provisioning as well
